Moved existing ES6 specs to specs folder.
