+++ app/javascript/spec/miq-component/helpers.spec.jsx
@@ -1,5 +1,5 @@
-import { define } from './registry.ts';
-import { cleanVirtualDom } from './helpers';
+import { define } from '../../miq-component/registry.ts';
+import { cleanVirtualDom } from '../../miq-component/helpers';
 
 describe('Helpers', () => {
   it('Should call instance destroy method if component mounting element is missing', () => {
