+++ app/javascript/spec/miq-redux/redux-router-actions.test.js
@@ -1,8 +1,8 @@
 import configureStore from 'redux-mock-store';
 import { connectRouter } from 'connected-react-router';
-import createReduxRoutingActions from './redux-router-actions';
-import createMiddlewares from './middleware';
-import { history } from '../miq-component/react-history.ts';
+import createReduxRoutingActions from '../../miq-redux/redux-router-actions';
+import createMiddlewares from '../../miq-redux/middleware';
+import { history } from '../../miq-component/react-history.ts';
 
 describe('Redux routing actions', () => {
   const mockStore = configureStore(createMiddlewares(history));
